refactor(retrofit): split retry logic from RetrofitExceptionHandler into BaseRetrofitExceptionHandler
to pave the way to share it with a new SpinnakerServerExceptionHandler class
