+++ orca-retrofit/src/test/groovy/com/netflix/spinnaker/orca/retrofit/exceptions/RetrofitExceptionHandlerSpec.groovy
@@ -17,22 +17,14 @@
 package com.netflix.spinnaker.orca.retrofit.exceptions
 
 import com.fasterxml.jackson.databind.ObjectMapper
-import retrofit.RestAdapter
 import retrofit.RetrofitError
-import retrofit.client.Client
 import retrofit.client.Response
 import retrofit.converter.JacksonConverter
-import retrofit.http.*
 import retrofit.mime.TypedByteArray
-import retrofit.mime.TypedString
 import spock.lang.Specification
 import spock.lang.Subject
 import spock.lang.Unroll
-import static java.net.HttpURLConnection.HTTP_BAD_GATEWAY
-import static java.net.HttpURLConnection.HTTP_UNAVAILABLE
 import static retrofit.RetrofitError.Kind.HTTP
-import static retrofit.RetrofitError.Kind.NETWORK
-import static retrofit.RetrofitError.Kind.UNEXPECTED
 import static retrofit.RetrofitError.httpError
 
 class RetrofitExceptionHandlerSpec extends Specification {
@@ -65,6 +57,7 @@ class RetrofitExceptionHandlerSpec extends Specification {
       !shouldRetry
       exceptionType == "RetrofitError"
       operation == stepName
+      details.kind == HTTP
       details.url == url
       details.status == status
       details.error == error
@@ -96,6 +89,7 @@ class RetrofitExceptionHandlerSpec extends Specification {
       !shouldRetry
       exceptionType == "RetrofitError"
       operation == stepName
+      details.kind == HTTP
       details.url == url
       details.status == status
       details.error == error
@@ -111,166 +105,5 @@ class RetrofitExceptionHandlerSpec extends Specification {
     error = reason
     rootException = "java.lang.RuntimeException"
     message = "Something bad happened"
-  }
-
-  private interface DummyRetrofitApi {
-    @GET("/whatever")
-    Response get()
-
-    @HEAD("/whatever")
-    Response head()
-
-    @POST("/whatever")
-    Response post(@Body String data)
-
-    @PUT("/whatever")
-    Response put()
-
-    @PATCH("/whatever")
-    Response patch(@Body String data)
-
-    @DELETE("/whatever")
-    Response delete()
-
-    @GET("/whatever/{stuff}")
-    Response getWithArg(@Path("stuff") String stuff)
-  }
-
-  def "should not retry an internal error"() {
-    given:
-    def client = Stub(Client) {
-      execute(_) >> { throw new IllegalArgumentException("Path parameter is null") }
-    }
-
-    and:
-    def api = new RestAdapter.Builder()
-      .setEndpoint("http://localhost:1337")
-      .setClient(client)
-      .build()
-      .create(DummyRetrofitApi)
-
-    when:
-    def ex = expectingException {
-      api.getWithArg(null)
-    }
-
-    then:
-    with(handler.handle("whatever", ex)) {
-      details.kind == UNEXPECTED
-      !shouldRetry
-    }
-  }
-
-  @Unroll
-  def "should not retry a network error on a #httpMethod request"() {
-    given:
-    def client = Stub(Client) {
-      execute(_) >> { throw new IOException("network error") }
-    }
-
-    and:
-    def api = new RestAdapter.Builder()
-      .setEndpoint("http://localhost:1337")
-      .setClient(client)
-      .build()
-      .create(DummyRetrofitApi)
-
-    when:
-    def ex = expectingException {
-      api."$methodName"("whatever")
-    }
-
-    then:
-    with(handler.handle("whatever", ex)) {
-      details.kind == NETWORK
-      !shouldRetry
-    }
-
-    where:
-    httpMethod << ["POST", "PATCH"]
-    methodName = httpMethod.toLowerCase()
-  }
-
-  @Unroll
-  def "should retry a network error on a #httpMethod request"() {
-    given:
-    def client = Stub(Client) {
-      execute(_) >> { throw new IOException("network error") }
-    }
-
-    and:
-    def api = new RestAdapter.Builder()
-      .setEndpoint("http://localhost:1337")
-      .setClient(client)
-      .build()
-      .create(DummyRetrofitApi)
-
-    when:
-    def ex = expectingException {
-      api."$methodName"()
-    }
-
-    then:
-    with(handler.handle("whatever", ex)) {
-      details.kind == NETWORK
-      shouldRetry
-    }
-
-    where:
-    httpMethod << ["GET", "HEAD", "DELETE", "PUT"]
-    methodName = httpMethod.toLowerCase()
-  }
-
-  @Unroll
-  def "shouldRetry=#expectedShouldRetry an HTTP #statusCode on a #httpMethod request"() {
-    given:
-    def client = Stub(Client) {
-      execute(_) >> new Response("http://localhost:1337", statusCode, "bad gateway", [], new TypedString(""))
-    }
-
-    and:
-    def api = new RestAdapter.Builder()
-      .setEndpoint("http://localhost:1337")
-      .setClient(client)
-      .build()
-      .create(DummyRetrofitApi)
-
-    and:
-    def ex = expectingException {
-      if (httpMethod in ["POST", "PATCH"]) {
-        api."$methodName"("body")
-      } else {
-        api."$methodName"()
-      }
-    }
-
-    expect:
-    with(handler.handle("whatever", ex)) {
-      details.kind == HTTP
-      shouldRetry == expectedShouldRetry
-    }
-
-    where:
-    httpMethod | statusCode       || expectedShouldRetry
-    "POST"     | HTTP_BAD_GATEWAY || false
-    "PATCH"    | HTTP_BAD_GATEWAY || false
-    "POST"     | HTTP_UNAVAILABLE || true
-
-    "GET"      | HTTP_BAD_GATEWAY || true
-    "HEAD"     | HTTP_BAD_GATEWAY || true
-    "DELETE"   | HTTP_BAD_GATEWAY || true
-    "PUT"      | HTTP_BAD_GATEWAY || true
-
-
-    methodName = httpMethod.toLowerCase()
-  }
-
-  private static RetrofitError expectingException(Closure action) {
-    try {
-      action()
-      throw new IllegalStateException("Closure did not throw an exception")
-    } catch (RetrofitError e) {
-      return e
-    }
   }
 }
